+++ logic/asyncTipGeneratorsLogic.js
@@ -3,6 +3,8 @@ const LanguageDetector = require('languagedetect');
 
 const cld = require('cld');
 const LinkPreviewLogic = require('./linkPreviewLogic.js');
+const RetipLogic = require('./retipLogic');
+const NotificationLogic = require('./notificationLogic');
 const TipLogic = require('./tipLogic');
 const aeternity = require('../utils/aeternity.js');
 const Logger = require('../utils/logger');
@@ -65,6 +67,20 @@ module.exports = class AsyncTipGeneratorsLogic {
           acc[address] = await aeternity.getTokenMetaInfoCacheAccounts(address);
           return acc;
         }, Promise.resolve({}));
+    });
+  }
+
+  static async triggerFetchAllLocalRetips(tips) {
+    await lock.acquire('AsyncTipGeneratorsLogic.fetchAllLocalRetips', async () => {
+      const localRetips = await RetipLogic.fetchAllLocalRetips();
+      const remoteRetips = [...new Set(tips.map(tip => tip.retips.map(retip => ({ ...retip, parentTip: tip }))).flat())];
+      const remoteRetipIds = [...new Set(remoteRetips.map(retip => retip.id))];
+      const localRetipIds = [...new Set(localRetips.map(retip => retip.id))];
+
+      const difference = remoteRetipIds.filter(id => !localRetipIds.includes(id));
+
+      // Send appropriate notifications for new tips
+      await difference.asyncMap(id => NotificationLogic.handleNewRetip(remoteRetips.find(retip => retip.id === id)));
     });
   }
 };
